new Figures for calibrationcampaign.tex
calibrationcampaign.tex now also shows analysis results, hence the
figures. analysis.tex is now added to .gitignore. It is not necessary
anymore. Merging calibrationcampaign.tex and analysis.tex reduced number
of pages in the output PDF quite a bit (even though it might still be
too much). These commits are part of a version, that I gave to Jelena
(just focusing on the calibration campaign and analysis sections, which
I merged). Another commit is coming.